chore: adjust default GameSetup values
Enabling takeItems makes the game more difficult and is an obtrusive setting. Such settings should be disabled by default. It also results in the lore feature – which is cool – being hidden.

The timer is distracting and takes up screen space. With player timer times being added to item lore, it isn't necessary to see the timer constantly by default.